+++ CHANGELOG.md
@@ -122,6 +122,7 @@
 
 ### Fixed
 
+* nostr: adj. `NostrConnectURI` de/serialization according to NIP46 ([Yuki Kishimoto])
 * lmdb: add missing commit method call in `Store::delete` ([Yuki Kishimoto])
 * lmdb: fix unit tests ([Yuki Kishimoto])
 * lmdb: fix `Store::save_event` issues ([Yuki Kishimoto])